Use `url_to_socket_addrs` to resolve addresses.
