+++ nym-vpn-core/crates/nym-gateway-directory/src/error.rs
@@ -24,18 +24,6 @@ pub enum Error {
     #[error(transparent)]
     VpnApiClientError(#[from] nym_vpn_api_client::error::VpnApiClientError),
 
-    #[error("failed to resolve gateway hostname: {hostname}")]
-    FailedToDnsResolveGateway {
-        hostname: String,
-        source: nym_http_api_client::HickoryDnsError,
-    },
-
-    #[error("resolved hostname {0} but no IP address found")]
-    ResolvedHostnameButNoIp(String),
-
-    #[error("timed out while attempting to resolve hostname: {hostname}")]
-    HostnameResolutionTimeout { hostname: String },
-
     #[error("failed to lookup described gateways")]
     FailedToLookupDescribedGateways(#[source] nym_validator_client::ValidatorClientError),
 
@@ -84,9 +72,6 @@ pub enum Error {
 
     #[error("failed to lookup gateway ip for gateway {0}")]
     FailedToLookupIp(String),
-
-    #[error("the url {url} doesn't parse to a host and/or a port: {reason}")]
-    UrlError { url: url::Url, reason: String },
 
     #[error("the provided gateway information is malformed")]
     MalformedGateway,
