+++ nym-vpn-core/crates/nym-gateway-directory/src/helpers.rs
@@ -1,70 +1,11 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::{
-    net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr},
-    time::Duration,
-};
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
-use nym_common::trace_err_chain;
-use nym_http_api_client::HickoryDnsResolver;
+use nym_vpn_api_client::url_to_socket_addr;
 
-use crate::{Config, Error, error::Result, gateway_client::ResolvedConfig};
-
-// be generous with the resolution timeout
-const HOSTNAME_RESOLUTION_TIMEOUT: Duration = Duration::from_secs(10);
-
-async fn try_resolve_hostname(hostname: &str) -> Result<Vec<IpAddr>> {
-    tracing::debug!("Trying to resolve hostname: {hostname}");
-    let resolver = HickoryDnsResolver::default();
-
-    let addrs =
-        match tokio::time::timeout(HOSTNAME_RESOLUTION_TIMEOUT, resolver.resolve_str(hostname))
-            .await
-        {
-            Ok(Ok(addrs)) => addrs,
-            Ok(Err(err)) => {
-                trace_err_chain!(err, "Failed to resolve hostname");
-                return Err(Error::FailedToDnsResolveGateway {
-                    hostname: hostname.to_string(),
-                    source: err,
-                });
-            }
-            Err(_timeout) => {
-                return Err(Error::HostnameResolutionTimeout {
-                    hostname: hostname.to_string(),
-                });
-            }
-        };
-
-    tracing::debug!("Resolved to: {addrs:?}");
-
-    let ips = addrs.iter().collect::<Vec<_>>();
-    if ips.is_empty() {
-        return Err(Error::ResolvedHostnameButNoIp(hostname.to_string()));
-    }
-
-    Ok(ips)
-}
-
-async fn url_to_socket_addr(unresolved_url: &url::Url) -> Result<Vec<SocketAddr>> {
-    let port = unresolved_url
-        .port_or_known_default()
-        .ok_or(Error::UrlError {
-            url: unresolved_url.clone(),
-            reason: "missing port".to_string(),
-        })?;
-    let hostname = unresolved_url.host_str().ok_or(Error::UrlError {
-        url: unresolved_url.clone(),
-        reason: "missing hostname".to_string(),
-    })?;
-
-    Ok(try_resolve_hostname(hostname)
-        .await?
-        .into_iter()
-        .map(|ip| SocketAddr::new(ip, port))
-        .collect())
-}
+use crate::{Config, error::Result, gateway_client::ResolvedConfig};
 
 pub async fn resolve_config(config: &Config) -> Result<ResolvedConfig> {
     let nyxd_socket_addrs = url_to_socket_addr(config.nyxd_url()).await?;
