+++ nym-vpn-core/crates/nym-vpn-api-client/Cargo.toml
@@ -18,6 +18,7 @@ bip39 = { workspace = true, features = ["zeroize"] }
 bs58.workspace = true
 chrono = { workspace = true, features = ["serde"] }
 itertools.workspace = true
+nym-common.workspace = true
 nym-compact-ecash.workspace = true
 nym-config.workspace = true
 nym-contracts-common.workspace = true
@@ -33,6 +34,7 @@ sha2.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
+tokio = { workspace = true, features = ["full"] }
 time = { workspace = true, features = [
     "serde-human-readable",
     "serde-well-known",
@@ -42,7 +44,6 @@ url.workspace = true
 zeroize.workspace = true
 
 [dev-dependencies]
-tokio = { workspace = true, features = ["full"] }
 bip39 = { workspace = true, features = ["zeroize"] }
 
 [features]
