+++ nym-vpn-core/crates/nym-vpn-api-client/src/error.rs
@@ -122,6 +122,24 @@ pub enum VpnApiClientError {
 
     #[error("failed to post account")]
     PostAccount(#[source] Box<HttpClientError>),
+
+    #[error("failed to resolve gateway hostname: {hostname}")]
+    FailedToDnsResolveGateway {
+        hostname: String,
+        source: nym_http_api_client::HickoryDnsError,
+    },
+
+    #[error("resolved hostname {0} but no IP address found")]
+    ResolvedHostnameButNoIp(String),
+
+    #[error("timed out while attempting to resolve hostname: {hostname}")]
+    HostnameResolutionTimeout { hostname: String },
+
+    #[error("the url {url} doesn't parse to a host and/or a port: {reason}")]
+    UrlError { url: url::Url, reason: String },
+
+    #[error("the url {url} is invalid")]
+    InvalidUrl { url: String },
 }
 
 pub type Result<T, E = VpnApiClientError> = std::result::Result<T, E>;
