+++ nym-vpn-core/crates/nym-vpn-api-client/src/lib.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 pub mod error;
+pub mod helpers;
 pub mod request;
 pub mod response;
 pub mod types;
@@ -13,4 +14,5 @@ mod network_compatibility;
 mod routes;
 
 pub use client::{ResolverOverrides, VpnApiClient};
+pub use helpers::{str_to_socket_addr, url_to_socket_addr};
 pub use network_compatibility::NetworkCompatibility;
